Version for IP8 tested on macOS Catalina and Big Sur

+++ src/StringTokenizer.cpp
@@ -1,8 +1,16 @@
+/*
+*  StringTokenizer.cpp
+*  SOCKIT
+*
+*  Created by andrew on 20/06/12.
+*  Copyright 2012 __MyCompanyName__. All rights reserved.
+*
+*/
 #include "XOPStandardHeaders.h"
 #include "StringTokenizer.h"
 
 
-void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, int szDELIMITERS)
+void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, size_t szDELIMITERS)
 {
     // Skip delimiters at beginning.
 	string str((const char*)STR, szStr);
