+++ src/StringTokenizer.h
@@ -1,6 +1,6 @@
 /*
  *  StringTokenizer.h
- *  iPeek
+ *  SOCKIT
  *
  *  Created by andrew on 18/05/08.
  *  Copyright 2008 __MyCompanyName__. All rights reserved.
@@ -13,5 +13,5 @@
 using namespace std;
 
 /* in StringTokenizer.cpp */
-void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, int szDELIMITERS);
+void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, size_t szDELIMITERS);
